# lusevokteren-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FjordVind Lusevokteren API
  description: |
    REST API for sea lice monitoring in Norwegian aquaculture.

    ## Authentication
    Most endpoints require authentication via JWT Bearer token.

    ```
    Authorization: Bearer <your-token>
    ```

    ## Demo Mode
    In development, demo tokens are available:
    - `demo_token_admin` - Admin access
    - `demo_token_driftsleder` - Operations manager
    - `demo_token_røkter` - Fish farmer

  version: 1.1.0
  contact:
    name: FjordVind Support
    email: support@fjordvind.no
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.fjordvind.no
    description: Production server
...
paths:
...
  /api/samples:
    get:
      summary: List samples
      tags: [Samples]
      security:
        - bearerAuth: []
      parameters:
        - name: merd_id
          in: query
          schema:
            type: string
            format: uuid
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of samples
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  samples:
                    type: array
                    items:
                      $ref: '#/components/schemas/Sample'

    post:
      summary: Create sample
      tags: [Samples]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSampleRequest'
      responses:
        '201':
          description: Sample created
        '400':
          description: Validation error
  /api/samples/{id}:
    get:
      summary: Get sample by ID
      tags: [Samples]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sample data
        '404':
          description: Sample not found

  # ========== TREATMENTS ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Sample:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sample_id:
          type: string
        merd_id:
          type: string
          format: uuid
        dato:
          type: string
          format: date
        antall_fisk:
          type: integer
        temperatur:
          type: number
        fish_observations:
          type: array
          items:
            $ref: '#/components/schemas/FishObservation'
    FishObservation:
      type: object
      properties:
        fish_id:
          type: string
        voksne_hunnlus:
          type: integer
        bevegelige_lus:
          type: integer
        fastsittende_lus:
          type: integer
    CreateSampleRequest:
      type: object
      required:
        - merd_id
        - dato
        - antall_fisk
      properties:
        merd_id:
          type: string
          format: uuid
        dato:
          type: string
          format: date
        antall_fisk:
          type: integer
          minimum: 1
        temperatur:
          type: number
        notat:
          type: string
        fish_observations:
          type: array
          items:
            type: object
            properties:
              voksne_hunnlus:
                type: integer
              bevegelige_lus:
                type: integer
              fastsittende_lus:
                type: integer
...